common 워크스페이스 제작
