+++ backend/package.json
@@ -30,6 +30,7 @@
     "@nestjs/passport": "^11.0.5",
     "@nestjs/platform-express": "^11.0.1",
     "@nestjs/platform-socket.io": "^11.0.17",
+    "@nestjs/schedule": "^6.0.0",
     "@nestjs/websockets": "^11.0.17",
     "@prisma/client": "^6.6.0",
     "@types/bcrypt": "^5.0.2",
@@ -39,14 +40,14 @@
     "box2d-wasm": "^7.0.0",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.1",
+    "lodash": "^4.17.21",
     "passport": "^0.7.0",
     "passport-jwt": "^4.0.1",
     "passport-local": "^1.0.0",
     "reflect-metadata": "^0.2.2",
     "rxjs": "^7.8.1",
     "strip-ansi": "^6.0.1",
-    "uuid": "^11.1.0",
-    "@nestjs/schedule": "^6.0.0"
+    "uuid": "^11.1.0"
   },
   "devDependencies": {
     "@nestjs/cli": "^11.0.0",
