+++ common/src/types/MapEntity.type.ts
