+++ package.json
@@ -6,7 +6,8 @@
   "workspaces": {
     "packages": [
       "frontend",
-      "backend"
+      "backend",
+      "common"
     ]
   },
   "engines": {
@@ -19,6 +20,7 @@
     "@swc/core": "^1.10.7",
     "@types/express": "^5.0.0",
     "@types/jest": "^29.5.14",
+    "@types/lodash": "^4.17.17",
     "@types/node": "^22.10.7",
     "@types/supertest": "^6.0.2",
     "@typescript-eslint/eslint-plugin": "^7.18.0",
@@ -45,7 +47,7 @@
   },
   "scripts": {
     "dev": "concurrently \"yarn frontend:dev\" \"yarn backend:dev\"",
-    "build": "yarn workspace frontend build && yarn workspace backend build",
+    "build": "yarn workspace common build && yarn workspace frontend build && yarn workspace backend build",
     "lint": "yarn workspace frontend lint && yarn workspace backend lint",
     "frontend:lint": "yarn workspace frontend lint",
     "frontend:dev": "yarn workspace frontend start",
@@ -54,7 +56,7 @@
     "backend:build": "yarn workspace backend build",
     "backend:lint": "yarn workspace backend lint",
     "prisma:generate": "yarn workspace backend prisma generate",
-    "postinstall": "yarn workspace backend prisma generate"
+    "postinstall": "yarn workspace common build && yarn workspace backend prisma generate"
   },
   "dependencies": {
     "caniuse-lite": "^1.0.30001713"
